add bpm parsing for single timing point maps and bump version

+++ src/beatmap/mod.rs
@@ -62,7 +62,7 @@ pub struct ParserBeatmap
     pub tags:              Vec<String>,
     pub combo_colors:      Vec<Color>,
     pub max_combo:         u32,
-    pub bpm:               Option<u32>,
+    pub bpm:               Option<f32>,
     pub circles:           u32,
     pub sliders:           u32,
     pub spinners:          u32,
